refactor: replace app_config_kv.h with app_base_config.h
Consolidate KV config helpers and filesystem/memory macros from
tool_files.h into a single app_base_config.h header. Update all
references across the codebase.
