+++ memory/session_manager.h
@@ -14,6 +14,7 @@
 #define __SESSION_MANAGER_H__
 
 #include "tuya_cloud_types.h"
+#include "app_base_config.h"
 #include <stddef.h>
 #include <stdint.h>
